Extract a common dialog component for normal and error flows
This will be useful to avoid visual jank from switching dialogs when an
error is encountered on the backend. Also, it removes some duplication,
and separates concerns.
